+++ pyaud/messages.py
@@ -5,7 +5,6 @@ pyaud.messages
 AUDIT_PASSED = "Success: All checks have passed"
 AUDIT_FAILED = "Failed: One or more checks have failed"
 AUDIT_RUNNING = "running the following plugins"
-COMMAND_NOT_FOUND = "{command}: Command not found"
 TYPE_ERROR = "can only register one of the following: {valid}; not {invalid}"
 FAILED = "Failed: returned non-zero exit status {returncode}"
 NAME_CONFLICT_ERROR = "plugin name conflict at {plugin}: '{name}'"